refactor: StockPriceUpdateController内のfetch処理をmoduleに切り出す
